+ | Returning an account for a token
