+++ backend/ponk/utils.py
@@ -1,0 +1,11 @@
+def get_selected_skin_url(username):
+    user = User.objects.get(username=username)
+    if user.selected_skin == "":
+        return random.choice(DEFAULT_SKINS)
+
+    skin = None
+    for item in skins:
+        if item["name"] == user.selected_skin:
+            skin = item
+
+    return random.choice(skin["images"])
